fix circular dependency for spring boot 2.6.x (#190) (#191)
Split autoconfiguration and generic bean configuraiton classes. This way BeelineAutoconfig can autowire BeelineHandlerInterceptor declared in either BeelineConfig or somewhere else. Existing tests already cover the change.
