+++ CONTRIBUTORS
@@ -3,4 +3,5 @@ honeycombio/beeline-java contributors:
 Christian Jank
 Thomas O'Hagan
 Peter Royal
+Giorgi Modebadze
 [Andy Coates](https://github.com/big-andy-coates)
